requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
cssselect2==0.7.0
Django==5.0.6
django-ckeditor-5==0.2.13
django-cors-headers==4.3.1
django-markdownx==4.0.7
django-rest-framework==0.1.0
djangorestframework==3.15.1
html5lib==1.1
idna==3.7
lxml==5.2.2
Markdown==3.6
oscrypto==1.3.0
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
pypdf==4.2.0
pypng==0.20220715.0
python-bidi==0.4.2
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.32.2
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.1
webencodings==0.5.1
wheel==0.43.0
xhtml2pdf==0.2.15
